+++ src/01-one/lexicon/methods/index.js
@@ -1,6 +1,6 @@
 import { toPresent, toPast, toFuture, toImperative, toGerund, isPerfective, getAspectPair, all as allVerb } from './verbs/conjugate.js'
 import { fromPresent, fromPast, fromImperative } from './verbs/to-root.js'
-import { toPlural, toSingular, decline, guessGender } from './nouns/inflect.js'
+import { toPlural, toSingular, decline, declinePlural, guessGender, guessAnimate } from './nouns/inflect.js'
 import { toMasculine, toFeminine, toNeuter, toPlural as toPluralAdj, toComparative, toSuperlative } from './adjectives/agree.js'
 
 export default {
@@ -21,7 +21,9 @@ export default {
     toPlural,
     toSingular,
     decline,
+    declinePlural,
     guessGender,
+    guessAnimate,
   },
   adjective: {
     toMasculine,